+++ projects/journey-maps-client/src/lib/services/map/map-routes.service.ts
@@ -2,10 +2,20 @@ import {Injectable} from '@angular/core';
 import {MapService} from './map.service';
 import {MapRouteService} from './map-route.service';
 import {Map as MaplibreMap} from 'maplibre-gl';
-
+import {ROUTE_ID_PROPERTY_NAME, SELECTED_PROPERTY_NAME} from './events/route-utils';
+import {SelectableFeatureCollection} from '../../journey-maps-client.interfaces';
 
 @Injectable({providedIn: 'root'})
 export class MapRoutesService {
+
+  static allRouteLayers: string[] = [
+    'rokas-route',
+    'rokas-route-gen0',
+    'rokas-route-gen1',
+    'rokas-route-gen2',
+    'rokas-route-gen3',
+    'rokas-route-gen4'
+  ];
 
   constructor(
     private mapService: MapService,
@@ -13,7 +23,14 @@ export class MapRoutesService {
   ) {
   }
 
-  updateRoutes(map: MaplibreMap, routes: GeoJSON.FeatureCollection[] = [this.mapService.emptyFeatureCollection]): void {
+  updateRoutes(map: MaplibreMap, routes: SelectableFeatureCollection[] = [this.mapService.emptyFeatureCollection]): void {
+    routes.forEach((featureCollection, idx) => {
+      const id = featureCollection.id ?? `jmc-generated-${idx + 1}`;
+      for (const f of featureCollection.features) {
+        f.properties[ROUTE_ID_PROPERTY_NAME] = id;
+        f.properties[SELECTED_PROPERTY_NAME] = featureCollection.isSelected;
+      }
+    });
     this.mapRouteService.updateRoute(map, {
       type: 'FeatureCollection',
       // With ES2019 we can replace this with routes.flatMap(({features}) => features)