+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+* Added support for map events for MARKER, ROUTE and STATION feature types. Use configuration setting in `listenerOptions` to enable the following events:
+  * `selectedFeaturesChange`
+  * `featuresClick`
+  * `featuresHoverChange`
 * An optional 'home button' has been added to the controls, allowing the user to zoom all of Switzerland into view.
 * `rokas-popup` and `rokas-teaser` components are made available to J-M-C integrators.
 * The option to show the map control buttons in a smaller size.